+++ pyproject.toml
@@ -84,6 +84,8 @@ convention = "numpy"
     "D100", # Missing docstring in public module
     "D102", # Missing docstring in public method
     "D103", # Missing docstring in public function
+    "D104", # Missing docstring in public package
+    "D205", # 1 blank line required between summary line and description
 ]
 "*/__init__.py" = [
     "F401", # Unused import
