+++ tests/component/lightgbm/test_booster_reproducibility.py
@@ -22,24 +22,6 @@ def test_booster_reproducible_from_booster_string(diabetes_data, lgb_diabetes_mo
     np.testing.assert_array_equal(predictions, predictions_reproduced)
 
 
-def test_convert_to_editable_booster(lgb_diabetes_model):
-    booster_string = BoosterString(lgb_diabetes_model)
-
-    editable_booster = convert_booster_string_to_editable_booster(booster_string)
-
-    assert len(editable_booster.trees) == 10
-
-
-def test_convert_editable_booster_to_booster_string(lgb_diabetes_model):
-    booster_string = BoosterString(lgb_diabetes_model)
-
-    editable_booster = convert_booster_string_to_editable_booster(booster_string)
-
-    booster_string_b = editable_booster._to_booster_string()
-
-    assert type(booster_string_b) is BoosterString
-
-
 def test_booster_reproducible_from_editable_booster(diabetes_data, lgb_diabetes_model):
     predictions = lgb_diabetes_model.predict(diabetes_data["data"])
